Fix date and time handling (#1140)
Using `time.UnixMicro()` not a good idea for Snowflake since Snowflake
decides on its own whether to interpret a number as milliseconds,
microseconds and so on. For a date such as 1970-03-02 which is much
smaller than 2023-09-08 or something, Snowflake interpets it as
milliseconds, leading to wildly incorrect Date values on the target.
This PR removes the usage of Unix.Micro() for Snowflake AVRO

This PR also adds tests for time for QRep

Fixes #1127